requirements.txt
backcall==0.1.0
cloudpickle==1.2.2
cycler==0.10.0
decorator==4.4.1
imageio==2.6.1
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
jedi==0.15.1
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==3.0.2
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.5
python-dateutil==2.8.1
pyzmq==18.1.1
scipy==1.3.3
six==1.13.0
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
